+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Purpose locale and icon with lowercase purpose
 * Remove square color in tooltip for PieChart
 * Fix speed values
+* Timeline is correctly sorted by date
 
 ## 🔧 Tech